[MER-5798] Clean up adaptive rule reference fix
